--- src/gbserver/webhooks/event_storage.py
@@ -1,11 +1,0 @@
-"""Backward-compatible re-exports. Storage moved to gbserver.storage."""
-from gbserver.storage.sql.webhook_event_storage import (  # noqa: F401
-    SQLWebhookEventStorage,
-    SqliteWebhookEventStorage,
-    create_webhook_event_storage,
-)
-from gbserver.storage.webhook_event_storage import (  # noqa: F401
-    BaseWebhookEventStorage,
-    GB_WEBHOOK_EVENTS_TABLE_NAME,
-    IWebhookEventStorage,
-)
