--- src/gbserver/webhooks/sql_storage.py
@@ -1,6 +1,0 @@
-"""Backward-compatible re-exports. Storage moved to gbserver.storage."""
-from gbserver.storage.sql.webhook_subscription_storage import (  # noqa: F401
-    SQLWebhookStorage,
-    SqliteWebhookStorage,
-    create_webhook_storage,
-)
